fix: address Codex round 4 on PR #5 (CSS-parsing edge cases)
- border-bottom width: strip the colour before reading the width (so 'solid
  #111344' isn't read as width 111344), treat 'none' / '0' as borderless, and
  default a width-less style/colour to a medium 1pt. [Codex]
- Clamp signature-line width to 1-100% and avoid a zero-flex spacer for ~full
  widths. [Codex]
- _matchDiv returns the real closing-tag end so a spaced '</div >' advances
  correctly. [Codex]
- marginCm: .toDouble() after clamp (clamp returns num). [Codex]
- Centred footer honours showDate. [Codex]

+++ lib/models/print_profile.dart
@@ -200,8 +200,9 @@ class PrintProfile {
         confidentialLabel: json['confidentialLabel'] as String?,
         // Clamp to the editor's slider range so an imported value stays
         // representable/adjustable in the profile editor.
-        marginCm:
-            ((json['marginCm'] as num?)?.toDouble() ?? 2.0).clamp(1.0, 3.5),
+        marginCm: ((json['marginCm'] as num?)?.toDouble() ?? 2.0)
+            .clamp(1.0, 3.5)
+            .toDouble(),
         accentRule: json['accentRule'] as bool? ?? true,
         headingRule: json['headingRule'] as bool? ?? false,
         footerCentered: json['footerCentered'] as bool? ?? false,
